+++ carts/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CartConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'cart'
+    name = 'carts'
