+++ carts/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path, include
-from cart.views import AddCartView
+from carts.views import AddCartView
 
 urlpatterns = [
     path("addcart", AddCartView.as_view())
